Initial support for hurd
